!T (Vulkan) Fix instances of narrowing, either by using a more appropriate type/constant, or by static_cast when we can be confident of the result. Enable -Wtautological-compare, -Wc++11-narrowing, and -Wempty-body.
Copied from Perforce
 Change: 1685122

+++ Code/CryEngine/CryCommon/CryCore/Platform/LinuxSpecific.h
@@ -106,7 +106,7 @@ typedef const WCHAR* LPCUWSTR, * PCUWSTR;
 typedef int      BOOL;
 typedef int32_t  LONG;
 typedef uint32_t ULONG;
-typedef int      HRESULT;
+typedef long     HRESULT;
 typedef uint32_t HMONITOR;
 
 #if !_MSC_EXTENSIONS
